implement obj parser

+++ RayTracerChallengev3/src/Patterns/StripePattern.cpp
@@ -6,7 +6,7 @@ namespace RayTracer
 {
 	Color StripePattern::patternAt(const Point4& p) const
 	{
-		int x = (int)std::floorf(p.x);
+		int x = (int)std::floor(p.x);
 		if (x % 2 == 0)
 		{
 			return color1;
